+++ src/printer_model_scaml.ml
@@ -1346,6 +1346,15 @@ let pp_model fmt (model : model) =
           f x
 
 
+      (* voting *)
+
+      | Mtotalvotingpower -> pp_str fmt "totalvotingpower"
+
+      | Mvotingpower x ->
+        Format.fprintf fmt "votingpower (%a)"
+          f x
+
+
       (* constants *)
 
       | Mnow           -> pp_str fmt "Global.get_now ()"